requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
pydantic-settings==2.4.0
python-dotenv==1.0.1
